requirements.txt
streamlit
pandas
numpy
plotly
python-dotenv
pyrebase4
scikit-learn
requests
